requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Video processing and computer vision
numpy<2.0.0
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# Job queue and caching
redis==5.0.1

# File handling and validation
python-multipart==0.0.6
python-magic==0.4.27

# Data validation and serialization
pydantic==2.5.0

# HTTP client for testing
httpx==0.25.2

# Testing framework
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Additional utilities
aiofiles==23.2.1
jinja2==3.1.2